[SCL-2142] Add REST API IP Safelist  (#78)
The changes in this PR adds a section to developer documentation for an IP safelist for the REST API. We are in the process of migrating the load balancers between VPCs which involves changing their IP addresses. To address this and future changes we would like to maintain an IP address safelist in the developer docs to have a single source of truth and limit friction for future upgrades that might change these IP addresses.