Remove reference to old and new IP lists on REST API IP lPs page (#107)
A few months ago, the REST API IPs page was updated to remove 2 lists of old IP addresses (one for each of the service regions). The page text still had a reference to the old and new lists, so I have removed that in this PR as it is no longer the case.